fix(architecture): move InMemoryStreamingEventStore to EventSourcing namespace
Architecture convention (NamingConventionTests) requires *EventStore types
to live under Compendium.Infrastructure.EventSourcing. CI caught it on
PR #89's first run.
